pyproject.toml
[project]
name = "vbiji"
version = "0.1.0"
description = "与 AI 一起协作的知识库 CLI 工具"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Nicky Hu", email = "68409250@qq.com" }
]
keywords = ["ai", "knowledge-base", "cli", "pdf", "llm", "deepseek", "kimi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.12.0",
    "PyMuPDF>=1.24.0",
    "python-docx>=1.1.0",
    "loguru>=0.5.0",
    "rich>=13.0.0",
    "click>=8.0.0",
    "toml>=0.10.0",
    "httpx>=0.28.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "mypy>=1.8.0", "ruff>=0.1.0", "types-toml"]

[project.scripts]
vbiji = "vbiji.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vbiji"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "typer>=0.24.2",
]

[project.urls]
Homepage = "https://github.com/ppweek/vbiji"
Repository = "https://github.com/ppweek/vbiji"
Documentation = "https://github.com/ppweek/vbiji#readme"
Issues = "https://github.com/ppweek/vbiji/issues"
Changelog = "https://github.com/ppweek/vbiji/blob/main/CHANGELOG.md"
